+++ src/onegov.notice/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Fix default direction when ordering notices.
+  [msom]
+
 0.0.4 (2017-08-11)
 ~~~~~~~~~~~~~~~~~~~